Put Gruntfile.js on a diet
The old gruntfile was tricky to understand and had some superfluous
configurations (like CoffeeScript, for example). I stripped out
everything but the bare essentials so we have something to work on and
added some comments about the build process.

For simplicity's sake, the source files are no longer uglified during
the build process, though this is trivial to fix later on.