Removed all unit tests (OMG!)
This is more drastic than it sounds; all in all there were roughly 4
test cases written that didn't really test that much. There was,
however, quite a bit of complexity added to run those 4 tests making the
project harder to work with.

I removed the test cases (spec/tests.js) and the dependencies needed to
run those tests. Unit tests are good and we'll want to be writing lots
of them, so this of course doesn't mean the end of testing.