Split endev.js into lots of small files
Instead of everything being in the same file, every directive and
factory in the endev angular module is now in its own file; the src/
folder now contains the subfolders directives/ and factories/. This is
for two reasons: it should make the codebase easier to read, and unit
testing should be less painful / feel more natural.

The Gulp build process has been updated to reflect this. The Grunt build
process is now broken for the JS side of things, but is needeed at the
moment since I haven't gotten round to replicating the html2js task in
Gulp. Once that is done, the Grunt dependencies will be removed of
course.