Browserify all the things!
Oh my god, I can't believe we've got to this point. The build system now
uses browserify to let us require() modules (in the node sense) from
within the code.

Why is this good?

It's not. It's freaking amazing! It means we're almost* at the
point where the code is completely modularized, and we can start
breaking stuff without changing the behaviour of the rest of the
codebase.

* I say almost because the only reason this works at the moment is
because we explicitly attach some variables to the window object. Next
up is to go through the code and remove those dependencies.