Removed global regexes from endev module
The regexes that used to be defined in the endev module are now moved to
the submodules (directives, services etc.) that use them. This does
entail some duplication, and perhaps these should be put in their own
module later on, even though it seems kind of silly to have a special regex
module.