Exposed global endev variable after initialization
For the sake of some examples, a global endev variable containing the
endev angular module is now exposed. However, this can only be accessed
_after_ initialization has taken place; you it is not possible to
reference it in the endev code.

This is necessary if one is to be able to dynamically add
directives etc. in script tags in the HTML document, for example.