Fixed the paths of most example libraries
The examples have been enumerated and three things have been changed in
all of the ones that are in a working state:

 - Bootstrap URL now points to a CDN
 - Endev URL now points to ../dist/endev.full.js
 - External libraries are removed, now the full endev library is always
   loaded (with angular, underscore etc. bundled)

The Bootstrap URL had to be changed since there's no bower_components
folder (or anything else that has to do with bower for that matter)
anymore. In some, but not all, examples the endev URL used to point to
endevjs.org, which meant that any local changes weren't reflected in the
examples. And lastly, there is at the minute no endev build that does
not bundle all its dependencies, though that most probably will change
in the future.