`package.json` and `bower.json` in `dist/`
A `package.json` and `bower.json` are now generated as part of the build
process. The idea behind putting them in the `dest/` folder is that the
`dest/` folder should be the _only_ thing distributed via npm/bower;
there's no reason to download all the source code and depedencies if you
just want to use the module.

The nice thing about generating them (instead of hardcoding) is that we
can use a lot of the same info from the original package.json, meaning
less duplication and less chance of inconsistent versioning etc :)